pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "podlens"
version = "1.2.17"
description = "Intelligent Podcast & Youtube Transcription & Understanding AI Agent"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Dunyuan Zha", email = "henryzha@outlook.com"}
]
maintainers = [
    {name = "Dunyuan Zha", email = "henryzha@outlook.com"}
]
keywords = ["podcast", "transcription", "summary", "youtube", "apple podcast", "whisper", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "requests",
    "feedparser>=6.0.0",
    "python-dotenv",
    "google-generativeai",
    "beautifulsoup4",
    "tqdm",
    "mlx-whisper",
    "groq",
    "openai-whisper",
    "youtube-transcript-api",
    "yt-dlp",
    "schedule",
]

[project.scripts]
# 主要命令 / Main commands
podlens = "podlens.cli_en:main"
pod = "podlens.cli_ch:main"

# 自动化命令 / Automation commands (integrated into CLI)
autopodlens = "podlens.cli_en:main"
autopod = "podlens.cli_ch:main"

# 向后兼容命令 / Backward compatibility commands
podlens-zh = "podlens.cli_ch:main"
podlens-en = "podlens.cli_en:main"
podlens-ch = "podlens.cli_ch:main"

[project.urls]
Homepage = "https://github.com/512z/podlens/tree/main"
Repository = "https://github.com/512z/podlens/tree/main"


[tool.setuptools]
packages = ["podlens"]

[tool.setuptools.package-data]
podlens = ["*.py"] 